WiP Introduction of a CDFT workflow
