+++ NOTICE
@@ -1,1 +1,1 @@
-This product includes nano-CAT, software developed by Bas van Beek.
+This product includes nano-CAT, software developed by B. F. van Beek.
